backend/pyproject.toml
[tool.poetry]
name = "document-search-app"
version = "0.1.0"
description = "Text search for db"
authors = ["Cmil30 <calebmigosi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.87.0"
pytest = "^7.2.0"
black = "^22.10.0"
uvicorn = "^0.19.0"
elasticsearch = {extras = ["async"], version = "^8.5.0"}
python-dotenv = "^0.21.0"
asyncio = "^3.4.3"
elastic-apm = "^6.13.2"


[tool.poetry.group.dev.dependencies]
numpy = "^1.23.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
